requirements.txt
# ML & Data
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
scikit-learn==1.3.2

# Flask & Extensions
Flask==2.3.3
Flask-Cors==3.0.10
flask_sqlalchemy==3.0.5
flask_jwt_extended==4.4.4

# Auth & Security
bcrypt==4.0.1
passlib==1.7.4
PyJWT==2.8.0

# Database
pymongo==4.4.0
   